add 2018/day16 part 1, 2

+++ src/main/util/helpers.cpp
@@ -154,13 +154,15 @@ vector<int> digits_of(string s) {
 	return out;
 }
 
-vector<long> extract_nums_from(string s) {
+// extracts all tokens resembling numbers (positive and negative)
+// does not work with decimal numbers or scientific notation (e.g. 9e9)
+vector<int> extract_nums_from(string s) {
 	const regex digits_regex("(-?\\d+)+");
 
 	auto begin_it = sregex_iterator(all(s), digits_regex);
 	auto end_it = sregex_iterator();
 
-	vector<long> out;
+	vector<int> out;
 	for (auto it = begin_it; it != end_it; ++it) {
 		smatch match = *it;
 		out.push_back(stoi(match.str()));
